pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "SwiftWinds", email = "12981958+SwiftWinds@users.noreply.github.com"},
]
dependencies = [
    "discord.py~=1.7",
    "python-dotenv~=0.19",
]
requires-python = ">=3.8"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
